scalar argument passing works again
The algorithm for argument passing was modified to account for the
size of each individual argument, in order to handle arguments which
do not fit within a single general purpose register.

Work still needs to be done to test if composite argument passing
works. and if composite values can be accessed correctly.