Modified Operand to use smaller members, This had an effect on each place in the compiler which interacts with Operands.
the total size is reduced from ~16 bytes to 4 bytes to improve allocation size
of all structures which hold operands, such as Tuples, and Bytecode Instructions.

There is now a limit of
2^16 ssa locals per function,
2^16 total unique constants per translation unit,
2^16 total unique labels per translation unit.
and immediate values must be within the range of an i16.

The backend still needs to be extended in order to account
for integer values out of the range of an immediate. This
includes all binary and unary operations. This is simple
enough to accomplish, as long as the integer is within
2^64 it can be stored as a unique constant in the translation unit.