refactoring the end to end tests to use small self contained "test resource" source files, instead of hardcoding the file contents into strings in the tests themselves.
This is to aid in debugging, as we can start up an instance of the
compiler with the problematic source in question under a debugger,
without having to first create the source file.

additionally, should this project ever grow to a size where it has
CVE's or something like that, these self contained source files can
be named with the issue in question, and that can associate the
test proving the issue is resolved with the file showing the
one problematic code.

and if I ever get around to setting up a fuzzer, the input that
causes the fuzz test to fail can be turned into a file, and added
to this resource directory for the same reasons.

all-in-all this seems like a more robust and extensible testing
strategy than the previous hardcoding approach. It doesn't replace
unit testing obviously, but it does work for the end to end testing.