added x64/imr and x64/env to better organize the x64 directory.
The naming convention is deliberately the same as the base directory. include/imr and include/env. because they have the same purpose, one to provide the in memory representation and environment for the x86_64 codegen, and one to provide the same for the generic bytecode format of the compiler.

We could have name imr/ IR/ instead, it's a matter of preference.